+++ ICD.Connect.Misc.Yepkit/Properties/AssemblyInfo.cs
@@ -1,0 +1,7 @@
+using System.Reflection;
+
+[assembly: AssemblyTitle("ICD.Connect.Misc.Yepkit")]
+[assembly: AssemblyCompany("ICD Systems")]
+[assembly: AssemblyProduct("ICD.Connect.Misc.Yepkit")]
+[assembly: AssemblyCopyright("Copyright © ICD Systems 2020")]
+[assembly: AssemblyVersion("8.0.0.0")]